`VariableStructure`のファイル分割を実施
子クラスごとに分割した
